requirements.txt
openai
python-dotenv
flask
pytest
waitress
